requirements.txt
mysqlclient==2.2.7
Django==5.1.6
djangorestframework==3.15.2